requirements.txt
flask
mysql-connector-python
werkzeug
pytesseract
Pillow
opencv-python
